Github finnicking
This repository is a forked version from Ahmed's. Both versions contain versions of code that do not reflect local/extreme changes. I am taking local changes from local/extreme and simply copying/pasting them to this directory in order to have a relatively up to date fork. This fork will not track changes and updates will be managed locally for the time being.